+++ OggConverter/src/Forms/Form1.Designer.cs
@@ -70,6 +70,8 @@
             this.tabs = new System.Windows.Forms.TabControl();
             this.tabLog = new System.Windows.Forms.TabPage();
             this.tabDownload = new System.Windows.Forms.TabPage();
+            this.labProgress = new System.Windows.Forms.Label();
+            this.proYt = new System.Windows.Forms.ProgressBar();
             this.ytdlOutput = new System.Windows.Forms.TextBox();
             this.btnCancelDownload = new System.Windows.Forms.Button();
             this.label6 = new System.Windows.Forms.Label();
@@ -87,8 +89,12 @@
             this.contextDelete = new System.Windows.Forms.ToolStripMenuItem();
             this.contextMove = new System.Windows.Forms.ToolStripMenuItem();
             this.contextAll = new System.Windows.Forms.ToolStripMenuItem();
-            this.proYt = new System.Windows.Forms.ProgressBar();
-            this.labProgress = new System.Windows.Forms.Label();
+            this.tabRecycle = new System.Windows.Forms.TabPage();
+            this.trashList = new System.Windows.Forms.ListBox();
+            this.btnRecycleDelete = new System.Windows.Forms.Button();
+            this.btnRestore = new System.Windows.Forms.Button();
+            this.labRecycle = new System.Windows.Forms.Label();
+            this.btnEmptyAll = new System.Windows.Forms.Button();
             this.panel1.SuspendLayout();
             this.menu.SuspendLayout();
             this.dragDropPanel.SuspendLayout();
@@ -97,6 +103,7 @@
             this.tabDownload.SuspendLayout();
             this.tabMeta.SuspendLayout();
             this.songListContext.SuspendLayout();
+            this.tabRecycle.SuspendLayout();
             this.SuspendLayout();
             // 
             // logOutput
@@ -285,7 +292,7 @@
             this.btnDel.Name = "btnDel";
             this.btnDel.Size = new System.Drawing.Size(64, 28);
             this.btnDel.TabIndex = 13;
-            this.btnDel.Text = "Delete";
+            this.btnDel.Text = "Remove";
             this.btnDel.UseVisualStyleBackColor = true;
             this.btnDel.Click += new System.EventHandler(this.BtnDel_Click);
             // 
@@ -484,6 +491,7 @@
             this.tabs.Controls.Add(this.tabLog);
             this.tabs.Controls.Add(this.tabDownload);
             this.tabs.Controls.Add(this.tabMeta);
+            this.tabs.Controls.Add(this.tabRecycle);
             this.tabs.HotTrack = true;
             this.tabs.Location = new System.Drawing.Point(340, 33);
             this.tabs.Margin = new System.Windows.Forms.Padding(4);
@@ -526,6 +534,24 @@
             this.tabDownload.Size = new System.Drawing.Size(645, 405);
             this.tabDownload.TabIndex = 1;
             this.tabDownload.Text = "Download";
+            // 
+            // labProgress
+            // 
+            this.labProgress.AutoSize = true;
+            this.labProgress.Location = new System.Drawing.Point(8, 158);
+            this.labProgress.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.labProgress.Name = "labProgress";
+            this.labProgress.Size = new System.Drawing.Size(159, 17);
+            this.labProgress.TabIndex = 21;
+            this.labProgress.Text = "0% 0.00 KiB/s ETA 0:00";
+            // 
+            // proYt
+            // 
+            this.proYt.Location = new System.Drawing.Point(11, 126);
+            this.proYt.Margin = new System.Windows.Forms.Padding(4);
+            this.proYt.Name = "proYt";
+            this.proYt.Size = new System.Drawing.Size(621, 28);
+            this.proYt.TabIndex = 20;
             // 
             // ytdlOutput
             // 
@@ -700,23 +726,75 @@
             this.contextAll.Text = "Select All";
             this.contextAll.Click += new System.EventHandler(this.ContextAll_Click);
             // 
-            // proYt
+            // tabRecycle
             // 
-            this.proYt.Location = new System.Drawing.Point(11, 126);
-            this.proYt.Margin = new System.Windows.Forms.Padding(4);
-            this.proYt.Name = "proYt";
-            this.proYt.Size = new System.Drawing.Size(621, 28);
-            this.proYt.TabIndex = 20;
+            this.tabRecycle.Controls.Add(this.btnEmptyAll);
+            this.tabRecycle.Controls.Add(this.labRecycle);
+            this.tabRecycle.Controls.Add(this.btnRestore);
+            this.tabRecycle.Controls.Add(this.btnRecycleDelete);
+            this.tabRecycle.Controls.Add(this.trashList);
+            this.tabRecycle.Location = new System.Drawing.Point(4, 25);
+            this.tabRecycle.Name = "tabRecycle";
+            this.tabRecycle.Size = new System.Drawing.Size(645, 405);
+            this.tabRecycle.TabIndex = 3;
+            this.tabRecycle.Text = "Recycle Bin";
+            this.tabRecycle.UseVisualStyleBackColor = true;
             // 
-            // labProgress
+            // trashList
             // 
-            this.labProgress.AutoSize = true;
-            this.labProgress.Location = new System.Drawing.Point(8, 158);
-            this.labProgress.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
-            this.labProgress.Name = "labProgress";
-            this.labProgress.Size = new System.Drawing.Size(159, 17);
-            this.labProgress.TabIndex = 21;
-            this.labProgress.Text = "0% 0.00 KiB/s ETA 0:00";
+            this.trashList.FormattingEnabled = true;
+            this.trashList.HorizontalScrollbar = true;
+            this.trashList.ItemHeight = 16;
+            this.trashList.Location = new System.Drawing.Point(16, 25);
+            this.trashList.Margin = new System.Windows.Forms.Padding(4);
+            this.trashList.Name = "trashList";
+            this.trashList.SelectionMode = System.Windows.Forms.SelectionMode.MultiExtended;
+            this.trashList.Size = new System.Drawing.Size(613, 324);
+            this.trashList.Sorted = true;
+            this.trashList.TabIndex = 24;
+            // 
+            // btnRecycleDelete
+            // 
+            this.btnRecycleDelete.Location = new System.Drawing.Point(16, 357);
+            this.btnRecycleDelete.Margin = new System.Windows.Forms.Padding(4);
+            this.btnRecycleDelete.Name = "btnRecycleDelete";
+            this.btnRecycleDelete.Size = new System.Drawing.Size(132, 44);
+            this.btnRecycleDelete.TabIndex = 25;
+            this.btnRecycleDelete.Text = "Delete";
+            this.btnRecycleDelete.UseVisualStyleBackColor = true;
+            this.btnRecycleDelete.Click += new System.EventHandler(this.BtnRecycleDelete_Click);
+            // 
+            // btnRestore
+            // 
+            this.btnRestore.Location = new System.Drawing.Point(497, 358);
+            this.btnRestore.Margin = new System.Windows.Forms.Padding(4);
+            this.btnRestore.Name = "btnRestore";
+            this.btnRestore.Size = new System.Drawing.Size(132, 44);
+            this.btnRestore.TabIndex = 26;
+            this.btnRestore.Text = "Restore";
+            this.btnRestore.UseVisualStyleBackColor = true;
+            this.btnRestore.Click += new System.EventHandler(this.BtnRestore_Click);
+            // 
+            // labRecycle
+            // 
+            this.labRecycle.AutoSize = true;
+            this.labRecycle.Location = new System.Drawing.Point(13, 6);
+            this.labRecycle.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.labRecycle.Name = "labRecycle";
+            this.labRecycle.Size = new System.Drawing.Size(137, 17);
+            this.labRecycle.TabIndex = 27;
+            this.labRecycle.Text = "Recycle bin is empty";
+            // 
+            // btnEmptyAll
+            // 
+            this.btnEmptyAll.Location = new System.Drawing.Point(156, 357);
+            this.btnEmptyAll.Margin = new System.Windows.Forms.Padding(4);
+            this.btnEmptyAll.Name = "btnEmptyAll";
+            this.btnEmptyAll.Size = new System.Drawing.Size(132, 44);
+            this.btnEmptyAll.TabIndex = 28;
+            this.btnEmptyAll.Text = "Empty All";
+            this.btnEmptyAll.UseVisualStyleBackColor = true;
+            this.btnEmptyAll.Click += new System.EventHandler(this.BtnEmptyAll_Click);
             // 
             // Form1
             // 
@@ -761,6 +839,8 @@
             this.tabMeta.ResumeLayout(false);
             this.tabMeta.PerformLayout();
             this.songListContext.ResumeLayout(false);
+            this.tabRecycle.ResumeLayout(false);
+            this.tabRecycle.PerformLayout();
             this.ResumeLayout(false);
             this.PerformLayout();
 
@@ -826,6 +906,12 @@
         private System.Windows.Forms.TextBox ytdlOutput;
         private System.Windows.Forms.Label labProgress;
         private System.Windows.Forms.ProgressBar proYt;
+        private System.Windows.Forms.TabPage tabRecycle;
+        private System.Windows.Forms.Button btnRecycleDelete;
+        public System.Windows.Forms.ListBox trashList;
+        private System.Windows.Forms.Button btnRestore;
+        private System.Windows.Forms.Label labRecycle;
+        private System.Windows.Forms.Button btnEmptyAll;
     }
 }