+++ OggConverter/src/Music/Player.cs
@@ -37,6 +37,15 @@ namespace OggConverter
 
         public static bool IsBusy { get; set; }
 
+        public static int GetNewFileNumber(string folder)
+        {
+            int newNumber = 1;
+            for (int i = 1; File.Exists($"{Settings.GamePath}\\{folder}\\track{i}.ogg"); i++)
+                newNumber++;
+
+            return newNumber;
+        }
+
         /// <summary>
         /// Plays selected song using ffplay.exe
         /// </summary>
@@ -239,9 +248,7 @@ namespace OggConverter
 
             try
             {
-                int newNumber = 1;
-                for (int i = 1; File.Exists($"{Settings.GamePath}\\{destination}\\track{i}.ogg"); i++)
-                    newNumber++;
+                int newNumber = GetNewFileNumber(destination);
 
                 // Waiting for file to be free
                 while (!Utilities.IsFileReady($"{Settings.GamePath}\\{source}\\{fileName}.ogg")) { }
@@ -282,9 +289,7 @@ namespace OggConverter
             try
             {
                 // Getting a new name for cloned song
-                int newNumber = 1;
-                for (int i = 1; File.Exists($"{Settings.GamePath}\\{folder}\\track{newNumber}.ogg"); i++)
-                    newNumber++;
+                int newNumber = GetNewFileNumber("folder");
 
                 newName = $"track{newNumber}";
 
@@ -363,11 +368,108 @@ namespace OggConverter
         }
 
         /// <summary>
-        /// Deletes the songs from list.
+        /// Moves the file to "Recycle Bin" folder inside of folder param.
         /// </summary>
         /// <param name="folder"></param>
-        /// <param name="fileName"></param>
-        /// <param name="songName"></param>
+        /// <param name="files"></param>
+        public static void Remove(string folder, string[] files)
+        {
+            if (Utilities.IsToolBusy())
+            {
+                Form1.instance.Log(Localisation.Get("Program is busy."));
+                return;
+            }
+
+            try
+            {
+                Player.Stop();
+
+                foreach (string file in files)
+                {
+                    string filePath = $"{Settings.GamePath}\\{folder}\\{file}.ogg";
+                    if (File.Exists(filePath))
+                    {
+                        if (!Directory.Exists($"{Settings.GamePath}\\Recycle Bin"))
+                            Directory.CreateDirectory($"{Settings.GamePath}\\Recycle Bin");
+
+                        while (!Utilities.IsFileReady(filePath)) { }
+
+                        string name = MetaData.GetName(file.Split('.')[0]);
+                        File.Move(filePath, $"{Settings.GamePath}\\Recycle Bin\\{name}.ogg");
+                        MetaData.Remove(file);
+
+                        Logs.History(Localisation.Get("Moved '{0}' ({1}) from {2} to recycle bin", name, file, folder));
+                        Form1.instance.Log(Localisation.Get("Moved '{0}' ({1}) from {2} to recycle bin", name, file, folder));
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                ErrorMessage err = new ErrorMessage(ex);
+                err.ShowDialog();
+            }
+            finally
+            {
+                if (Settings.AutoSort)
+                    Player.Sort(folder);
+
+                Form1.instance.UpdateSongList();
+            }
+        }
+
+        /// <summary>
+        /// Restores files from the recycle bin
+        /// </summary>
+        /// <param name="folder"></param>
+        /// <param name="files"></param>
+        public static void Restore(string folder, string[] files)
+        {
+            if (Utilities.IsToolBusy())
+            {
+                Form1.instance.Log(Localisation.Get("Program is busy."));
+                return;
+            }
+
+            if (!Directory.Exists($"{Settings.GamePath}\\Recycle Bin"))
+                return;
+            
+            try
+            {
+                foreach (string file in files)
+                {
+                    string filePath = $"{Settings.GamePath}\\Recycle Bin\\{file}.ogg";
+                    if (File.Exists(filePath))
+                    {
+                        while (!Utilities.IsFileReady(filePath)) { }
+
+                        string newFileName = $"track{GetNewFileNumber(folder)}";
+                        File.Move(filePath, $"{Settings.GamePath}\\{folder}\\{newFileName}.ogg");
+                        MetaData.AddOrEdit(newFileName, file);
+                        
+                        Logs.History(Localisation.Get("Restored '{0}' from the recycle bin", file));
+                        Form1.instance.Log(Localisation.Get("Restored '{0}' from the recycle bin", file));
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                ErrorMessage err = new ErrorMessage(ex);
+                err.ShowDialog();
+            }
+            finally
+            {
+                if (Settings.AutoSort)
+                    Player.Sort(folder);
+
+                Form1.instance.UpdateSongList();
+            }
+        }
+
+        /// <summary>
+        /// Deletes the songs permamently. Includes file and MetaData entry
+        /// </summary>
+        /// <param name="folder">Working folder</param>
+        /// <param name="files">Array of file</param>
         public static void Delete(string folder, string[] files)
         {
             if (Utilities.IsToolBusy())
@@ -428,6 +530,42 @@ namespace OggConverter
                 if (Settings.AutoSort)
                     Player.Sort(folder);
 
+                Form1.instance.UpdateSongList();
+            }
+        }
+
+        /// <summary>
+        /// Empties the recycle bin
+        /// </summary>
+        public static void DeleteAll()
+        {
+            if (Utilities.IsToolBusy())
+            {
+                Form1.instance.Log(Localisation.Get("Program is busy."));
+                return;
+            }
+
+            try
+            {
+                DirectoryInfo di = new DirectoryInfo($"{Settings.GamePath}\\Recycle Bin");
+                FileInfo[] files = di.GetFiles("*.ogg");
+
+                string message = Localisation.Get("Are you sure you want to delete all {0} files?", files.Length);
+                DialogResult dl = MessageBox.Show(message, Localisation.Get("Question"), MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+
+                if (dl == DialogResult.Yes)
+                {
+                    foreach (var file in files)
+                        file.Delete();
+                }
+            }
+            catch (Exception ex)
+            {
+                ErrorMessage err = new ErrorMessage(ex);
+                err.ShowDialog();
+            }
+            finally
+            {
                 Form1.instance.UpdateSongList();
             }
         }